refactor: Split StructureNbtManager and LargeStructureNbtManager into focused modules (Issue #3)
## Summary
Refactored two large manager classes to improve maintainability and comply with
the 800-line file guideline.

## Changes

### StructureNbtManager (1,313 → 235 lines, 82% reduction)
- Extracted shared utilities to `utils/StructureNbtUtils.kt` and `utils/NbtConversionUtils.kt`
- Split into focused operation modules:
  - `operations/StructureNbtCore.kt` - Core file operations (load, save, list, remove)
  - `operations/StructureNbtCapture.kt` - World capture and placement
  - `operations/StructureNbtBlocks.kt` - Block analysis and replacement
  - `operations/StructureNbtBackup.kt` - Backup and restore operations
- Main file now serves as a clean facade pattern

### LargeStructureNbtManager (1,100 → 195 lines, 82% reduction)
- Split into focused operation modules:
  - `large/LargeStructureNbtCore.kt` - Core operations (list, remove, getSize, blocksReplace)
  - `large/LargeStructureNbtCapture.kt` - Capture/place with grid splitting logic
  - `large/LargeStructureNbtBackup.kt` - Directory-based backup management
- Main file now serves as a clean facade pattern
- Reuses shared utilities from StructureNbtUtils

## Benefits
- ✅ Improved maintainability - clear separation of concerns
- ✅ Better testability - each module can be tested independently
- ✅ Extracted common code - reduced duplication between managers
- ✅ Preserved all functionality - no breaking changes
- ✅ Follows project guidelines - all files now < 800 lines

## Related
Addresses Issue #3: Reduce file sizes for large manager classes
